fix(generate-examples-index): produce the same hash everywhere
The support files use path hashes for unique identification. The
absolute path was used which resulted in CI tests failing because the
path was different on each computer. This should resolve the issue.

Also since it is the same for all URLs related to the same example it is
now generated only once for each example and reused rather than
regenerated for each URL again.