fix(rollup-conf): styles dir location
Due to recent breaking changes in Rollup PostCSS plugin that I
completely missed the styles dir has moved from the root to the bundle
files and is now 4 times in the packages (4 copies of the same thing).
This is not how it is documented and was never announced to the public
as a breaking change. This commit reverses it to the way it used to be
prior to the update. It also officially drops support for Rollup PostCSS
plugin (dev dep in our projects) version 2 since the way paths are
interpreted is different from the up to date version 3.