fix(babel): rewrite Browserlist query
Due to a weird quirk in the syntax of Browserlist both `or not dead` and
`and not dead` are intersections with the previos part of the query.
Using `or` here is very confusing as it implies that we support browsers
that match one (>1% usage) or the other (not dead) but in fact they have
to match both. It definitely confused me (the comment above the query is
a proof of it).

This changes nothing regarding the browsers we support as in this case
`(A && !B) === (A || !B)` so the list of supported and polyfilled
browsers stays exactly the same.