chore(package): replace opencollective by npm fund
We used this to notify people about the funding of our project but this
is now handled by npm natively via fund in the package and via npm fund.
Furthermore the dependency is no longer maintained (last update is
6 years ago).

Also, at least on my machine, it only seems to output:
```
> opencollective postinstall || exit 0

sh: line 1: opencollective: command not found
```
anyway.