perf(babel-preset)!: remove 7 years dead CSS plugin
BREAKING CHANGE: When importing CSS, it will no longer be possible to
get objects mapping class names to style in JS.

I didn't find anything that would break because of this. If it was used
somewhere, I'll have to change to some other package anyway because this
one has seen no releases in the past 7 years.